move appproject functions to the appproject manager package
