+++ principal/callbacks_test.go
@@ -363,236 +363,6 @@ func TestSyncAppProjectUpdatesToAgents(t *testing.T) {
 	}
 }
 
-func TestAgentSpecificAppProject(t *testing.T) {
-	tests := []struct {
-		name       string
-		appProject v1alpha1.AppProject
-		agent      string
-		want       v1alpha1.AppProject
-	}{
-		{
-			name: "filters destinations for matching agent",
-			appProject: v1alpha1.AppProject{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "test-project",
-					Namespace: "argocd",
-				},
-				Spec: v1alpha1.AppProjectSpec{
-					Destinations: []v1alpha1.ApplicationDestination{
-						{
-							Name:      "cluster-prod",
-							Namespace: "default",
-							Server:    "https://prod-cluster.example.com",
-						},
-						{
-							Name:      "cluster-dev",
-							Namespace: "staging",
-							Server:    "https://dev-cluster.example.com",
-						},
-					},
-					SourceNamespaces: []string{"argocd", "apps"},
-					Roles: []v1alpha1.ProjectRole{
-						{
-							Name: "admin",
-							Policies: []string{
-								"p, proj:test-project:admin, applications, *, test-project/*, allow",
-							},
-						},
-					},
-				},
-			},
-			agent: "cluster-prod",
-			want: v1alpha1.AppProject{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "test-project",
-					Namespace: "argocd",
-				},
-				Spec: v1alpha1.AppProjectSpec{
-					Destinations: []v1alpha1.ApplicationDestination{
-						{
-							Name:      "in-cluster",
-							Namespace: "default",
-							Server:    "https://kubernetes.default.svc",
-						},
-					},
-					SourceNamespaces: []string{"cluster-prod"},
-					Roles:            []v1alpha1.ProjectRole{},
-				},
-			},
-		},
-		{
-			name: "matches multiple destinations with glob pattern",
-			appProject: v1alpha1.AppProject{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "multi-dest-project",
-					Namespace: "argocd",
-				},
-				Spec: v1alpha1.AppProjectSpec{
-					Destinations: []v1alpha1.ApplicationDestination{
-						{
-							Name:      "prod-cluster-1",
-							Namespace: "app1",
-							Server:    "https://prod1.example.com",
-						},
-						{
-							Name:      "prod-cluster-1",
-							Namespace: "app2",
-							Server:    "https://prod2.example.com",
-						},
-						{
-							Name:      "dev-cluster",
-							Namespace: "app3",
-							Server:    "https://dev.example.com",
-						},
-					},
-					SourceNamespaces: []string{"*"},
-					Roles: []v1alpha1.ProjectRole{
-						{Name: "viewer"},
-						{Name: "admin"},
-					},
-				},
-			},
-			agent: "prod-cluster-1",
-			want: v1alpha1.AppProject{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "multi-dest-project",
-					Namespace: "argocd",
-				},
-				Spec: v1alpha1.AppProjectSpec{
-					Destinations: []v1alpha1.ApplicationDestination{
-						{
-							Name:      "in-cluster",
-							Namespace: "app1",
-							Server:    "https://kubernetes.default.svc",
-						},
-						{
-							Name:      "in-cluster",
-							Namespace: "app2",
-							Server:    "https://kubernetes.default.svc",
-						},
-					},
-					SourceNamespaces: []string{"prod-cluster-1"},
-					Roles:            []v1alpha1.ProjectRole{},
-				},
-			},
-		},
-		{
-			name: "no matching destinations",
-			appProject: v1alpha1.AppProject{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "no-match-project",
-					Namespace: "argocd",
-				},
-				Spec: v1alpha1.AppProjectSpec{
-					Destinations: []v1alpha1.ApplicationDestination{
-						{
-							Name:      "staging-cluster",
-							Namespace: "default",
-							Server:    "https://staging.example.com",
-						},
-						{
-							Name:      "test-cluster",
-							Namespace: "testing",
-							Server:    "https://test.example.com",
-						},
-					},
-					SourceNamespaces: []string{"argocd", "staging"},
-					Roles: []v1alpha1.ProjectRole{
-						{
-							Name: "operator",
-							Policies: []string{
-								"p, proj:no-match-project:operator, applications, get, *, allow",
-							},
-						},
-					},
-				},
-			},
-			agent: "prod-cluster",
-			want: v1alpha1.AppProject{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "no-match-project",
-					Namespace: "argocd",
-				},
-				Spec: v1alpha1.AppProjectSpec{
-					Destinations:     []v1alpha1.ApplicationDestination{},
-					SourceNamespaces: []string{"prod-cluster"},
-					Roles:            []v1alpha1.ProjectRole{},
-				},
-			},
-		},
-		{
-			name: "empty destinations and roles",
-			appProject: v1alpha1.AppProject{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "empty-project",
-					Namespace: "argocd",
-				},
-				Spec: v1alpha1.AppProjectSpec{
-					Destinations:     []v1alpha1.ApplicationDestination{},
-					SourceNamespaces: []string{"original-namespace"},
-					Roles:            []v1alpha1.ProjectRole{},
-				},
-			},
-			agent: "test-agent",
-			want: v1alpha1.AppProject{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "empty-project",
-					Namespace: "argocd",
-				},
-				Spec: v1alpha1.AppProjectSpec{
-					Destinations:     []v1alpha1.ApplicationDestination{},
-					SourceNamespaces: []string{"test-agent"},
-					Roles:            []v1alpha1.ProjectRole{},
-				},
-			},
-		},
-		{
-			name: "agent name matches wildcard destination",
-			appProject: v1alpha1.AppProject{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "wildcard-project",
-					Namespace: "argocd",
-				},
-				Spec: v1alpha1.AppProjectSpec{
-					Destinations: []v1alpha1.ApplicationDestination{
-						{
-							Name:      "*",
-							Namespace: "default",
-						},
-					},
-					SourceNamespaces: []string{"*"},
-					Roles:            []v1alpha1.ProjectRole{},
-				},
-			},
-			agent: "test-agent",
-			want: v1alpha1.AppProject{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "wildcard-project",
-					Namespace: "argocd",
-				},
-				Spec: v1alpha1.AppProjectSpec{
-					Destinations: []v1alpha1.ApplicationDestination{
-						{
-							Name:      "in-cluster",
-							Server:    "https://kubernetes.default.svc",
-							Namespace: "default",
-						},
-					},
-					SourceNamespaces: []string{"test-agent"},
-					Roles:            []v1alpha1.ProjectRole{},
-				},
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got := AgentSpecificAppProject(tt.appProject, tt.agent)
-			assert.Equal(t, tt.want, got)
-		})
-	}
-}
-
 func TestIsResourceFromAutonomousAgent(t *testing.T) {
 	tests := []struct {
 		name    string
@@ -647,170 +417,6 @@ func TestIsResourceFromAutonomousAgent(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			got := isResourceFromAutonomousAgent(&tt.project)
-			assert.Equal(t, tt.want, got)
-		})
-	}
-}
-
-func TestDoesAgentMatchWithProject(t *testing.T) {
-	tests := []struct {
-		name       string
-		agentName  string
-		appProject v1alpha1.AppProject
-		want       bool
-	}{
-		{
-			name:      "agent matches destination name",
-			agentName: "agent-prod",
-			appProject: v1alpha1.AppProject{
-				Spec: v1alpha1.AppProjectSpec{
-					Destinations: []v1alpha1.ApplicationDestination{
-						{Name: "agent-prod", Namespace: "default"},
-					},
-					SourceNamespaces: []string{"agent-prod"},
-				},
-			},
-			want: true,
-		},
-		{
-			name:      "empty destination name with wildcard server",
-			agentName: "any-agent",
-			appProject: v1alpha1.AppProject{
-				Spec: v1alpha1.AppProjectSpec{
-					Destinations: []v1alpha1.ApplicationDestination{
-						{Name: "", Server: "*", Namespace: "default"},
-					},
-					SourceNamespaces: []string{"any-agent"},
-				},
-			},
-			want: true,
-		},
-
-		{
-			name:      "agent matches destination name with wildcard",
-			agentName: "agent-staging",
-			appProject: v1alpha1.AppProject{
-				Spec: v1alpha1.AppProjectSpec{
-					Destinations: []v1alpha1.ApplicationDestination{
-						{Name: "agent-*", Namespace: "default"},
-					},
-					SourceNamespaces: []string{"agent-*"},
-				},
-			},
-			want: true,
-		},
-		{
-			name:      "deny pattern matches agent - should reject",
-			agentName: "prod-agent",
-			appProject: v1alpha1.AppProject{
-				Spec: v1alpha1.AppProjectSpec{
-					Destinations: []v1alpha1.ApplicationDestination{
-						{Name: "!prod-*", Namespace: "default"},
-					},
-					SourceNamespaces: []string{"*"},
-				},
-			},
-			want: false,
-		},
-		{
-			name:      "deny pattern with positive pattern - deny takes precedence",
-			agentName: "prod-agent",
-			appProject: v1alpha1.AppProject{
-				Spec: v1alpha1.AppProjectSpec{
-					Destinations: []v1alpha1.ApplicationDestination{
-						{Name: "*", Namespace: "default"},       // Allow all
-						{Name: "!prod-*", Namespace: "default"}, // Deny prod
-					},
-					SourceNamespaces: []string{"*"},
-				},
-			},
-			want: false,
-		},
-		{
-			name:      "deny pattern with positive pattern - allow others",
-			agentName: "dev-agent",
-			appProject: v1alpha1.AppProject{
-				Spec: v1alpha1.AppProjectSpec{
-					Destinations: []v1alpha1.ApplicationDestination{
-						{Name: "*", Namespace: "default"},       // Allow all
-						{Name: "!prod-*", Namespace: "default"}, // Deny prod
-					},
-					SourceNamespaces: []string{"*"},
-				},
-			},
-			want: true,
-		},
-
-		{
-			name:      "agent matches destination but not source namespace",
-			agentName: "agent-test",
-			appProject: v1alpha1.AppProject{
-				Spec: v1alpha1.AppProjectSpec{
-					Destinations: []v1alpha1.ApplicationDestination{
-						{Name: "agent-test", Namespace: "default"},
-					},
-					SourceNamespaces: []string{"different-namespace"},
-				},
-			},
-			want: false,
-		},
-		{
-			name:      "only deny patterns - no positive match",
-			agentName: "dev-agent",
-			appProject: v1alpha1.AppProject{
-				Spec: v1alpha1.AppProjectSpec{
-					Destinations: []v1alpha1.ApplicationDestination{
-						{Name: "!prod-*", Namespace: "default"},
-					},
-					SourceNamespaces: []string{"*"},
-				},
-			},
-			want: false, // No positive pattern to match
-		},
-		{
-			name:      "server URL with exact agent name match",
-			agentName: "prod-cluster-1",
-			appProject: v1alpha1.AppProject{
-				Spec: v1alpha1.AppProjectSpec{
-					Destinations: []v1alpha1.ApplicationDestination{
-						{Name: "", Server: "https://resource-proxy.example.com:8080?agentName=prod-cluster-1", Namespace: "default"},
-					},
-					SourceNamespaces: []string{"*"},
-				},
-			},
-			want: true, // Agent name matches extracted query parameter
-		},
-		{
-			name:      "server URL with agent name pattern",
-			agentName: "prod-cluster-2",
-			appProject: v1alpha1.AppProject{
-				Spec: v1alpha1.AppProjectSpec{
-					Destinations: []v1alpha1.ApplicationDestination{
-						{Name: "", Server: "https://resource-proxy.example.com:8080?agentName=prod-cluster-*", Namespace: "default"},
-					},
-					SourceNamespaces: []string{"*"},
-				},
-			},
-			want: true, // Agent name matches extracted pattern
-		},
-		{
-			name:      "server URL with non-matching agent name",
-			agentName: "dev-cluster-1",
-			appProject: v1alpha1.AppProject{
-				Spec: v1alpha1.AppProjectSpec{
-					Destinations: []v1alpha1.ApplicationDestination{
-						{Name: "", Server: "https://resource-proxy.example.com:8080?agentName=prod-cluster-1", Namespace: "default"},
-					},
-					SourceNamespaces: []string{"*"},
-				},
-			},
-			want: false, // Agent name doesn't match extracted query parameter
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got := doesAgentMatchWithProject(tt.agentName, tt.appProject)
 			assert.Equal(t, tt.want, got)
 		})
 	}
